+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # CHANGELOG
+
+## (2022-09-01)
+- added aws-classic-py-vpc-get-zones-awsx
 ## (2022-08-30)
 - added aws-classic-ts-vpc-get-zones-awsx
 ## (2022-08-22)